backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain and LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# DashScope SDK (for Wan2.6-t2v video generation)
dashscope==1.25.9

# OpenAI SDK (for ModelScope OpenAI-compatible endpoint)
openai>=1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Data Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Redis
redis==5.0.1

# Vector Store
faiss-cpu==1.7.4

# Template Engine
jinja2==3.1.2

# Logging
structlog==23.2.0

# File Processing
aiofiles==23.2.1

# API Validation
openapi-spec-validator==0.7.1
pyyaml
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dotenv==1.0.0
